pyproject.toml
[build-system]
requires = ["setuptools>=65.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thinkiplex"
description = "Download and organize Thinkific courses for Plex"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dynamic = ["version"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "pyyaml>=6.0",
    "inquirer>=3.1.3",
    "pathlib>=1.0.1",
    "ffmpeg-python>=0.2.0",
    "pydantic>=2.0.0",
    "concurrent-log-handler>=0.9.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "ruff>=0.0.40",
    "pre-commit>=3.0.0",
    "types-PyYAML>=6.0.12.12",
]

[project.scripts]
thinkiplex = "thinkiplex.main:main"

[tool.setuptools]
packages = ["thinkiplex"]

[tool.setuptools.dynamic]
version = {attr = "thinkiplex.__version__"}

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
target-version = "py38"
line-length = 100
select = ["E", "F", "B", "I"]